Add two more functions: onOperator and onOperatorAdded:
onOperator only executes if the last digit is a number,
and there is not a operator sign in the string.
With that, we can assign what's inside the button without having something like:

 "4+/5"

isOperatorAdded is an auxiliar function for onOperator,
it returns true if the string of the TextView have any operator sign,
and makes an exception for the minus sign,
allowing us to do math with negative numbers.

there is a "problem" with this. If we start the number with a minus,
the code allow us to do things like:

"-4+5/3*2-1"

it is not wrong mathematically, but just happens if we start with a minus...

Also changed the onClick function in the operator buttons to onOperator in the xml