Add a new function: onEqual:
this function split the text from TextView into separated values,
store them into an array, and makes a subtraction (so far) by converting
to Doubles and making the operation, then convert back to String and set the
TextView with the result.

For now, it just does subtractions.,
and if tried to do any other operation it'll crash the app.

Also change the onClick parameter to onEqual at the xml file.